chore(utils): optimize utils
