--- packages/components/_util/hasScrollBar.ts
@@ -1,3 +1,0 @@
-export default function hasScrollbar() {
-  return document.body.scrollHeight > (window.innerHeight || document.documentElement.clientHeight);
-}
