+++ packages/components/anchor/Anchor.tsx
@@ -5,10 +5,10 @@ import { StyledProps } from '../common';
 import { TdAnchorProps } from './type';
 import useConfig from '../hooks/useConfig';
 import forwardRefWithStatics from '../_util/forwardRefWithStatics';
-import { canUseDocument, getScrollContainer } from '../_util/dom';
+import { canUseDocument } from '../_util/dom';
 import Affix from '../affix';
 import { AnchorContext, Item } from './AnchorContext';
-import { getOffsetTop, getScroll, scrollTo, AnchorContainer } from './_util/dom';
+import { getOffsetTop, getScroll, scrollTo, AnchorContainer, getScrollContainer } from '../_util/scroll';
 import AnchorItem from './AnchorItem';
 import AnchorTarget from './AnchorTarget';
 import { anchorDefaultProps } from './defaultProps';
