--- packages/components/anchor/_util/dom.ts
@@ -1,84 +1,0 @@
-import raf from 'raf';
-import { easeInOutCubic, EasingFunction } from '../../_util/easing';
-
-export type AnchorContainer = HTMLElement | Window;
-
-export function getOffsetTop(target: HTMLElement, container: AnchorContainer) {
-  const { top } = target.getBoundingClientRect();
-  if (container === window) {
-    // 减去 document 的边框
-    return top - document.documentElement.clientTop;
-  }
-  return top - (container as HTMLElement).getBoundingClientRect().top;
-}
-
-/**
- * 返回是否 window 对象
- *
- * @export
- * @param {any} obj
- * @returns
- */
-function isWindow(obj: any) {
-  return obj && obj === obj.window;
-}
-
-type ScrollTarget = HTMLElement | Window | Document;
-
-/**
- * 获取滚动距离
- *
- * @export
- * @param {ScrollTarget} target
- * @param {boolean} isLeft true为获取scrollLeft, false为获取scrollTop
- * @returns {number}
- */
-export function getScroll(target: ScrollTarget, isLeft?: boolean): number {
-  // node 环境或者 target 为空
-  if (typeof window === 'undefined' || !target) {
-    return 0;
-  }
-  const method = isLeft ? 'scrollLeft' : 'scrollTop';
-  let result = 0;
-  if (isWindow(target)) {
-    result = (target as Window)[isLeft ? 'pageXOffset' : 'pageYOffset'];
-  } else if (target instanceof Document) {
-    result = target.documentElement[method];
-  } else if (target) {
-    result = (target as HTMLElement)[method];
-  }
-  return result;
-}
-
-interface ScrollTopOptions {
-  container?: ScrollTarget;
-  duration?: number;
-  easing?: EasingFunction;
-}
-export function scrollTo(target: number, opt: ScrollTopOptions) {
-  const { container = window, duration = 450, easing = easeInOutCubic } = opt;
-  const scrollTop = getScroll(container);
-  const startTime = Date.now();
-  return new Promise((res) => {
-    const fnc = () => {
-      const timestamp = Date.now();
-      const time = timestamp - startTime;
-      const nextScrollTop = easing(Math.min(time, duration), scrollTop, target, duration);
-      if (isWindow(container)) {
-        (container as Window).scrollTo(window.pageXOffset, nextScrollTop);
-      } else if (container instanceof HTMLDocument || container.constructor.name === 'HTMLDocument') {
-        (container as HTMLDocument).documentElement.scrollTop = nextScrollTop;
-      } else {
-        (container as HTMLElement).scrollTop = nextScrollTop;
-      }
-      if (time < duration) {
-        raf(fnc);
-      } else {
-        // 由于上面步骤设置了 scrollTop, 滚动事件可能未触发完毕
-        // 此时应该在下一帧再执行 res
-        raf(res);
-      }
-    };
-    raf(fnc);
-  });
-}
