+++ packages/components/auto-complete/OptionList.tsx
@@ -6,8 +6,8 @@ import log from '../../common/js/log';
 import { CommonClassNameType } from '../hooks/useCommonClassName';
 import { AutoCompleteOptionObj, TdAutoCompleteProps } from './type';
 import HighlightOption from './HighlightOption';
-import { on, off } from '../_util/dom';
 import { useLocaleReceiver } from '../locale/LocalReceiver';
+import { off, on } from '../_util/listener';
 
 export interface OptionsListProps {
   sizeClassNames: CommonClassNameType['sizeClassNames'];
