+++ packages/components/popup/Popup.tsx
@@ -17,7 +17,7 @@ import useWindowSize from '../hooks/useWindowSize';
 import { popupDefaultProps } from './defaultProps';
 import useDefaultProps from '../hooks/useDefaultProps';
 import useAttach from '../hooks/useAttach';
-import { getCssVarsValue } from '../_util/dom';
+import { getCssVarsValue } from '../_util/style';
 
 export interface PopupProps extends TdPopupProps {
   // 是否触发展开收起动画，内部下拉式组件使用
