+++ packages/components/popup/PopupPlugin.tsx
@@ -4,11 +4,12 @@ import { isString } from 'lodash-es';
 import classNames from 'classnames';
 import { CSSTransition } from 'react-transition-group';
 import { render, unmount } from '../_util/react-render';
-import { on, off, getAttach } from '../_util/dom';
+import { getAttach } from '../_util/dom';
 import { TNode } from '../common';
 import { TdPopupProps } from './type';
 import useDefaultProps from '../hooks/useDefaultProps';
 import { popupDefaultProps } from './defaultProps';
+import { off, on } from '../_util/listener';
 
 export interface PopupPluginApi {
   config: TdPopupProps;
